+++ sdk/metric/internal/aggregate/sum.go
@@ -8,13 +8,15 @@ import (
 	"time"
 
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/sdk/internal/x"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 )
 
 type sumValue[N int64 | float64] struct {
-	n     atomicCounter[N]
-	res   FilteredExemplarReservoir[N]
-	attrs attribute.Set
+	n         atomicCounter[N]
+	res       FilteredExemplarReservoir[N]
+	attrs     attribute.Set
+	startTime time.Time
 }
 
 type sumValueMap[N int64 | float64] struct {
@@ -29,9 +31,14 @@ func (s *sumValueMap[N]) measure(
 	droppedAttr []attribute.KeyValue,
 ) {
 	sv := s.values.LoadOrStoreAttr(fltrAttr, func(attr attribute.Set) any {
+		var startTime time.Time
+		if x.PerSeriesStartTimestamps.Enabled() {
+			startTime = now()
+		}
 		return &sumValue[N]{
-			res:   s.newRes(attr),
-			attrs: attr,
+			res:       s.newRes(attr),
+			attrs:     attr,
+			startTime: startTime,
 		}
 	}).(*sumValue[N])
 	sv.n.add(value)
@@ -160,12 +167,19 @@ func (s *cumulativeSum[N]) collect(
 	// current length for capacity.
 	dPts := reset(sData.DataPoints, 0, s.values.Len())
 
+	perSeriesStartTimeEnabled := x.PerSeriesStartTimestamps.Enabled()
+
 	var i int
 	s.values.Range(func(_, value any) bool {
 		val := value.(*sumValue[N])
+
+		startTime := s.start
+		if perSeriesStartTimeEnabled {
+			startTime = val.startTime
+		}
 		newPt := metricdata.DataPoint[N]{
 			Attributes: val.attrs,
-			StartTime:  s.start,
+			StartTime:  startTime,
 			Time:       t,
 			Value:      val.n.load(),
 		}